move TicketCreatedListener and Listener class to their own moduless
