+++ nats-test/src/events/ticket-created-listener.ts
@@ -1,0 +1,13 @@
+import { Message } from "node-nats-streaming";
+import { Listener } from "./base-listener";
+
+export class TicketCreatedListener extends Listener {
+  subject = "ticket:created";
+  queueGroupName = "payments-service";
+
+  onMessage(data: any, msg: Message) {
+    console.log("Event data!", data);
+
+    msg.ack();
+  }
+}
