+++ nats-test/src/listener.ts
@@ -1,5 +1,7 @@
-import nats, { Message, Stan } from "node-nats-streaming";
+import nats from "node-nats-streaming";
 import { randomBytes } from "crypto";
+
+import { TicketCreatedListener } from "./events/ticket-created-listener";
 
 console.clear();
 
@@ -15,77 +17,8 @@ stan.on("connect", () => {
     process.exit();
   });
 
-  const options = stan
-    .subscriptionOptions()
-    .setManualAckMode(true)
-    .setDeliverAllAvailable()
-    .setDurableName("orders-service");
-  const subscription = stan.subscribe(
-    "ticket:created",
-    "orders-service-queue-group",
-    options
-  );
-
-  subscription.on("message", (msg: Message) => {
-    const data = msg.getData();
-    if (typeof data === "string") {
-      console.log(
-        `Received event number ${msg.getSequence()}, with data: 
-          ${data}
-        }`
-      );
-    }
-
-    msg.ack();
-  });
+  new TicketCreatedListener(stan).listen();
 });
 
 process.on("SIGINT", () => stan.close());
 process.on("SIGTERM", () => stan.close());
-
-abstract class Listener {
-  abstract subject: string;
-  abstract queueGroupName: string;
-  abstract onMessage(data: any, msg: Message): void;
-
-  private client: Stan;
-  protected ackWait = 5 * 1000;
-
-  constructor(client: Stan) {
-    this.client = client;
-  }
-
-  subscriptionOptions() {
-    return this.client
-      .subscriptionOptions()
-      .setDeliverAllAvailable()
-      .setManualAckMode(true)
-      .setAckWait(this.ackWait)
-      .setDurableName(this.queueGroupName);
-  }
-
-  listen() {
-    const subscription = this.client.subscribe(
-      this.subject,
-      this.queueGroupName,
-      this.subscriptionOptions()
-    );
-
-    subscription.on("message", (msg: Message) => {
-      console.log(
-        `Message receieved: ${this.subject} / ${this.queueGroupName}`
-      );
-
-      const parsedData = this.parseMessage(msg);
-      this.onMessage(parsedData, msg);
-    });
-  }
-
-  parseMessage(msg: Message) {
-    const data = msg.getData();
-
-    return typeof data === "string"
-      ? JSON.parse(data)
-      : JSON.parse(data.toString("utf8"));
-  }
-}
